refactor(api): extract helpers, probes, and OpenAPI from app.py
Split 3077-line monolith into focused modules:
- _helpers.py: utility classes and functions (~345 lines)
- _probes.py: HTTP/WS endpoint health probes (~145 lines)
- _openapi.py: OpenAPI spec builder (~180 lines)

app.py retains RuntimeAPIContext + RuntimeAPIServer (~2500 lines).
TYPE_CHECKING guard prevents circular imports for _openapi.
All 1429 tests pass, mypy strict clean.
